fixes for copilot review
